Added entropy. Fixed mathematical error in template collection. Fixed active percentage bug.

+++ configs/dlgn_conv_config.py
@@ -15,7 +15,7 @@ class HardReLU_F(torch.autograd.Function):
     def forward(ctx, inputs):
         ctx.save_for_backward(inputs)  # save input for backward pass
         device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
-        values = torch.tensor([0.], device=device)
+        values = torch.tensor([0], dtype=inputs.dtype, device=device)
         retval = torch.heaviside(inputs, values)
         return retval
